move singledispatchmethod impl to data_manipulation so it can be used by interpreter and program context
